cli: move the RPC client code to a separate file
Release note: None
