+++ pkg/cli/debug_list_files.go
@@ -46,13 +46,11 @@ func runDebugListFiles(cmd *cobra.Command, _ []string) error {
 	defer cancel()
 
 	// Connect to the node pointed to in the command line.
-	conn, _, finish, err := getClientGRPCConn(ctx, serverCfg)
+	status, finish, err := getStatusClient(ctx, serverCfg)
 	if err != nil {
 		return err
 	}
 	defer finish()
-
-	status := serverpb.NewStatusClient(conn)
 
 	// Retrieve the details for the head node.
 	firstNodeDetails, err := status.Details(ctx, &serverpb.DetailsRequest{NodeId: "local"})
